+++ web-application-server-gradle/src/main/java/db/DataBase.java
@@ -1,8 +1,10 @@
 package db;
 
-import com.google.common.collect.Maps;
 import java.util.Collection;
 import java.util.Map;
+
+import com.google.common.collect.Maps;
+
 import model.User;
 
 public class DataBase {
