+++ web-application-server-gradle/src/main/java/util/IOUtils.java
@@ -1,12 +1,14 @@
-package utils;
+package util;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 
 public class IOUtils {
     /**
-     * @param BufferedReader는 Request Body를 시작하는 시점이어야
-     * @param contentLength는  Request Header의 Content-Length 값이다.
+     * @param BufferedReader는
+     *            Request Body를 시작하는 시점이어야
+     * @param contentLength는
+     *            Request Header의 Content-Length 값이다.
      * @return
      * @throws IOException
      */
