--- web-application-server-gradle/src/main/java/utils/enums/HttpMethod.java
@@ -1,5 +1,0 @@
-package utils.enums;
-
-public enum HttpMethod {
-    GET, HEAD, POST, PUT, DELETE, CONNECT, OPTIONS, TRACE, PATCH;
-}
