--- web-application-server-gradle/src/main/java/utils/enums/Series.java
@@ -1,5 +1,0 @@
-package utils.enums;
-
-public enum Series {
-    CLIENT_ERROR, SUCCESSFUL, REDIRECTION, INFORMATIONAL, SERVER_ERROR
-}
