+++ web-application-server-gradle/src/test/java/util/IOUtilsTest.java
@@ -1,4 +1,4 @@
-package utils;
+package util;
 
 import java.io.BufferedReader;
 import java.io.StringReader;
